+++ src/shared/components/AdjustStakeModal.js
@@ -116,13 +116,7 @@ const Note = styled.div(({ theme }) => ({
 })
 export default class AdjustStakeModal extends React.Component {
   render() {
-    const {
-      total,
-      handleSubmit,
-      submitting,
-      currentStake,
-      change,
-    } = this.props;
+    const { total, handleSubmit, submitting, change } = this.props;
     return (
       <Modal
         style={{ maxWidth: 600 }}
@@ -175,22 +169,9 @@ export default class AdjustStakeModal extends React.Component {
               >
                 {__('Cancel')}
               </Button>
-              <Field
-                name="stake"
-                component={({ input }) => (
-                  <Button
-                    skin="primary"
-                    type="submit"
-                    disabled={
-                      submitting ||
-                      // using == instead of === because input.value can be either string or number
-                      input.value == currentStake
-                    }
-                  >
-                    {__('Set stake amount')}
-                  </Button>
-                )}
-              />
+              <Button skin="primary" type="submit" disabled={submitting}>
+                {__('Set stake amount')}
+              </Button>
             </div>
           </form>
         </Modal.Body>