+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Disallowed logging in to your sigchain before the sigchain creation is confirmed on the blockchain
 - You can now use Escape key to close most pop-up modals
 - Unconfirmed transactions are now dimmed on Transactions page
+- Allow setting stake amount to the current amount (for the case you want to undo the change you made)
 
 #### Fixes